Don't error if boomark `page' is nil
`saveplace-pdf-view-save-to-alist' errors with a wrong-type-argument `numberp'
`nil' error if called in the context of a PDF whose
`pdf-view-bookmark-make-record' has a nil `page' due to an incorrect comparison:
`=' errors if the provided arguments are not numbers, which means that the `eq'
nil check must happen first. Comparison against `nil' should be done using
`null', so use that.