Make the package compatible with Emacs 30
There were some changes in the saveplace package in Emacs 30.

It now expects that if the Value in a Key-Value pair in
save-place-alist is a list, then it should probably be a string that
contains a path to a file that can be navigated to using dired.  It
will then either expand or abbreviate the path in that list based on
the setting `save-place-abbreviate-file-names`.

The problem is that it assumes that this value is always a string and
applies #'expand-file-name or #'abbreviate-file-name on it, causing
saveplace-pdf-view to break.

As a workaround, we wrap the bookmark entry in a vector so that
saveplace will skip the step that causes the breakage.